requirements.txt
django==5.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
django-filter==25.1 